created routes folder and task file. also modified main file to separate routes from backend
